requirements.txt
django
django-import-export
openpyxl
django-modeltranslation
requests
django-recaptcha==3.0.0
Pillow